requirements.txt
pandas
flask
numpy
pymongo
python-dotenv
PyYAML
-e . 